Fix user role is reset upon login. Resolve #26
== Note from Issue #26 ==
Currently when logging into wordpress using CAS with LDAP enabled the
user's roll is always set to the hard coded value of
"$defaultEmployeeUserrole" or "$defaultStudentUserrole" depending which
role the user matches.

These values should be used the first time the user logs in, but if the
user role is changed by an WordPress administrator the plugin should
not modify the user's role.