Fix #40 - LDAP user/pass not saving
LDAP username & password settings did not seem to be saving to
the database. Added those two fields to the retrieval of
settings (they were missing), but didn't find the save code
before it started working. Jonathan set me up with dev access
to the database at that time, so I'm thinking that something
happened with the DB permissions.